refactor: split workflow session tests
